Fix: wrong alignment check with monsters wishing for an artifact.
Noticed that the alignment check performed when a monster wishes for an
artifact was wrong. Using mon->malign, that's not checking a monsters
alignment, it's used as a value adjustment comparing the players
alignment vs the monster they did a thing to. So swap it out with the
proper function call, so monsters will wish for a properly co-aligned
artifact that they can actually use.