Fix: improper handling of worn/wielded objects that break.
This was first noticed when a the helm of brilliance had its base
material changed to glass (NetHack 3.7 change). Worn pieces of armor can
deflect attacks, and can sometimes break depending on their material
(glass, dark elven gear made of adamantine in the light), or if their
forged quality is inferior. What was happening - worn enchanted helm of
brilliance that wasn't 'fixed' was destroyed after blocking an attack,
but the int/wis bonus was still present. As of this commit, instead of
just calling setworn() to be NULL, we call the proper function to
correctly remove the destroyed worn piece of armor, which in turn will
remove any special effects it was giving.

While here, noticed that weapon handling could be improved (call
uwepgone() instead of just 'unweapon = TRUE'), and add handling for the
offweapon while twoweaponing (I tested this with a dark elven weapon in
the offhand while in the light; when it broke, the game crashed on the
next attack since the offhand weapon hadn't been properly cleared).