goodpos() improvements (open air terrain).
The way goodpos() function handled placing a monster or the player over
open air needed some improvement. I also increased the try count for
finding a good position when placing a randomly generated monster by a
factor of 10.

While here, I noticed that when a flag was set to nothing in goodpos(),
it was inconsistent throughout the code (0 should be 0L), so I addressed
that as well.