Fix: monsters breaking boulders fixes.
The big fix - don't rely on mspec_used (that's for monster
spellcasting), use a new int made specifically for boulder breaking.
Timeout values are unchanged.

Also in this commit - any quest leader can break boulders, not just
*your* quest leader. I also noticed that some monster values were not
being reset properly upon revival, and mbreakboulder needed to be a part
of that, so I added in those values as well (see zap.c).