Fix: feedback regarding the Bell of Opening when the quest leader demands return of the quest artifact.
The text originally said that the (paraphrasing) 'curse on the silver
bell has been lifted' when asking for the return of the quest artifact.
That's not true though, because if the quest artifact is not returned,
the quest leader gets angry, and the Bell of Opening is still 'cursed'.
So I've edited the text here slightly for each quest leader, states that
they're ready to lift the curse, but haven't actually done it yet.