Player riding a steed can now feed their steed by throwing food down at them.
Before, the only thing you could throw down to your steed while riding
it and not have it hit the floor was a potion (primarily for uncursing
your saddle with holy water). With this commit, you can now feed your
steed without having to dismount first. Just throw a food item down at
them, and if they're hungry, or if they view the food item as a treat,
they'll catch and eat it.

The main reason for this commit - it was very difficult to cure your
steed of zombie illness with a eucalyptus leaf. Because virtually all
steeds are flagged as an animal, they won't pick up and use a leaf if
sick like an intelligent monster would; you have to throw it to them
like a pet treat. By the time you dismount and get in-line with your
steed to throw a leaf at them, it could sometimes be too late.

Some code formatting here as well.