Crafting traps, part 3: additional trap types (bolt/dart/spear).
This commit adds bolt, dart, and spear traps to the fold. While here, I
did some changes with the recipe struct, makes it a bit easier to
understand.

Debating on whether to allow component enchantment to pass through.
Meaning if you use a stack of 10 +3 arrows for an arrow trap, the arrows
in the trap will also be +3. I doubt that would assist with a to-hit
bonus, but the extra damage should apply. Untrapping a random ammo-type
trap can produce enchanted ammo, so this seems appropriate. Keep an eye
out for that commit.