Crafting traps, part 4: beartraps and land mines.
This one was simple, just add the crafting recipes for beartraps and
land mines, the rest of the code for how they behave is already in
place.

Using a grappling hook to craft a beartrap seems appropriate. As does
using sling bullets for a land mine (the assumption is that the Rogue's
trap kit has the explosive components needed to finish the job). The
sling bullets will need to be made of their default material (iron) for
the process to work, as land mines are also iron and cannot be made of
any other material.