Crafting traps, part 5: additional trap types (sleeping gas/fire/ice/teleportation/magic/anti-magic/polymorph/magic beam).
This commit completes the remaining traps I wanted to be able to
crafted. The majority of these traps require a specific type of wand as
a component to make it. Sleeping gas trap requires a wand of sleeping,
fire trap requires a wand of fire, and so forth. The rule for wand
components - the wand used must have a minimum of six charges, otherwise
the trap build will fail. Magic beam traps require a wand of magic
missile as the component - because of this, crafted beam traps will
always fire magic missiles when triggered, as opposed to a random ray
type when found in the wild.

The one trap here that doesn't use a wand as a component is a magic
trap. These require three potions of gain energy instead. Normally,
monsters were mostly immune to magic traps, as there was only a 1 in 21
chance a monster would trigger it, and if it did, it'd act like a fire
trap. To make these traps worth crafting, I made it so there's a 50/50
chance of a monster setting one of these off, and if they do, the effect
could either be a magical explosion, or behave like a fire trap.

Now that all the craftable traps have been created, and their behavior
pretty much dialed in, I want to focus on making trap crafting/setting
an actual skill that can be trained. Some of these traps should be
easier to craft and set than others, and not every role should have the
training necessary to do it, especially for the more advanced traps like
teleportation and polymorph. I should adjust the base costs for the more
complex traps as well.