Crafting traps, part 6: successfully creating/setting traps odds.
Initially, the plan was to create a new skill the could be trained, and
that skill would be used to determing whether a trap could be
successfully created and set. I decided against that, since players of
all roles can already set beartraps and land mines. I thought it'd be
better to make trap crafting/setting more of a percentage chance of
success based on your dexterity, and your role will be a factor as well.

Crafting traps: if you're not a Rogue or a Ranger, there's a chance of
not using a trap kit correctly, and checks against your dexterity. The
lower your dexterity, the greater the chance of failing to use your trap
kit (having a dexterity of 20 or greater will never fail). Failure uses
up a charge on the trap kit.

Setting traps: this is a bit more involved. Dexterity is a factor here
again (the higher it is, the better), as is your role and your
experience level. Being a Rogue has the best odds of successfully
setting a trap (dexterity plus current experience level times 5); being
a Ranger has good odds as well, but not as good as a Rogue (dexterity
plus current experience level times 2). All other roles, there is no
multiplier, but the odds of successfully setting a trap still increase
as you level up. If setting a trap fails, there's another check against
your dexterity, and if that check fails, the trap you were trying to set
is destroyed (having a dexterity of 15 or greater, trap will never fall
apart).

I also added an encumberance check for crafting and setting traps - if
you're stressed or greater, neither will work.