Player gets the credit for monsters killed in a trap they made.
Extra incentive to make and set traps. If your trap kills a monster, you
get the experience for it. You can also take the blame if your trap
kills a pet, but luckily if you set a trap within sight of your pet,
they know about the trap and will avoid it.