Fix: monsters could loot locked containers without opening them if the container was being carried.
A player recently encountered a bug that has existed for more than six
years, but thankfully this bug is very rare. It's probably happened
before, but the circumstances surrounding this recent occurrence (castle
wand of wishing inside a crystal chest) got everyone's attention.

What happened: player had a tame chameleon. Pet turned into a winter
wolf, and picked up the crystal chest with the wand of wishing still
inside of it (locked). Before the pet dropped the chest, it turned into
a genetic engineer (humanoid, has hands) which immediately removed the
contents of the crystal chest without unlocking or opening it. It then
proceeded to use up the wand of wishing.

Going way way back to when EvilHack was brand new, and I was busy adding
code from GruntHack to EvilHack. One of the things ported over was a
monsters ability to loot containers, either on the ground or a container
in their inventory. mloot_container() handles containers that are on the
floor, but precheck() had code added to it that allowed monsters to
remove objects from a container they were carrying. The code added to
precheck() had no guards whatsoever for locked/empty/missing containers,
cursed bags of holding, and so on. Monsters don't normally pick up
things like boxes or chests, but they will if there's an object inside
of it that they want and they're strong enough to do so. This bug has
probably expressed itself before, but not quite like this.

So, with this commit, guards have been put in place for locked
containers and the like, and a few other improvements as well, such as
removing a corpse from a carried ice box.

There's a lot of room for improvement here. An intelligent monster could
have various objects in their carried bag that would be useful, but they
won't remove/use them like they would if they were looting a chest on
the ground. Can save this for a later date, for now just fixing the bug
will suffice.