Fix: player riding a steed can now feed their steed by throwing food down at them, take two.
Commit 408fe0c allowed the player riding a steed to feed it without
dismounting by simply throwing the food object down at their steed, as
if they were throwing it on the ground. This worked fine for food
objects the steed wanted to actually eat. If an object was thrown that
the steed didn't want or couldn't eat, that object disappeared
permanently, and could then cause a crash on game exit. This commit
fixes that bug, now uneaten objects will just land on the floor.