Adjust certain feedback when forging; update The Blacksmith's Cookbook.
This commit addresses issue #183 - certain feedback when forging could
be misleading, implying that any object made of stone (mineral) could be
forged. That bit was added when sling bullets were included a s forge
recipe (3 rocks plus a dart). The feedback when forged has been adjusted
so as not to be misleading. The Blacksmith's Cookbook has also been
updated in this regard.

While here, I added two more chapters to the Cookbook to cover
object forge quality.