Fix: illithid race players could still use 'psionic wave' if polymorphed into a non-illithid type monster.
The whole psionics piece is built around the premise that illithids,
mind flayers, etc have this massive ultra-complex brain structure that
allows them to use psionic abilities. I had handling for this already
when an illithid infidel crowned and transformed into a demon, but I
neglected to do this if an illithid player polymorphed into something
that wasn't illithid-like.