Accessibility - showdamage, showmondamage options.
The showdamage option comes from NetHack 3.7 commit dd37c53 (slightly
modified). This allows the player to set the option to show the damage
they've taken from an attacking monster or other source. Is off by
default, can be set in-game or set in the player's config file
(OPTIONS=showdamage).

The showmondamage option is something I decided to create to simplify
debugging damage being dealt to a monster via melee. This is also off by
default, but can be set in-game or by placing OPTIONS=showmondamage in
your config. This option is only available/works in wizmode. The routine
damage_mon() is currently the only place in the code that uses
showmondamage(), but that covers a lot of bases. This would need to be
expanded to cover every way a monster could be damaged by the player.