Allow all monsters that can wear a saddle or barding to have a chance to spawn wearing them.
Previously, only domestic monsters (ponies, horses, etc) had a chance to
spawn wearing a saddle or barding. This commit gets rid of the
is_domestic() requirement. Because of this, the odds of a monster that
can wear a saddle spawning with a saddle are decreased, as there will be
several more monster types that can now potentially spawn wearing one.