Fix: player races that have intrinsic flying could still fly when polymorphed into a non-flying monster.
Those being Illithids (level 12 or greater) and Vampires (level 7 or
greater). The fix is to remove intrinsic flying (FROMRACE) when polyd
into a non-flying form, but then give it back when rehumanize() is
called. Put a check in here to make sure those races were at the proper
experience level when their intrinsic flying was restored.

This feels a bit hacky, but it works. There may be a better way to do
this.