Fix: potions and forge dipping.
It was recently noticed that dipping potions into a forge did nothing to
them. Exception to that was a potion of oil would become lit, but
otherwise nothing. The expectation is that they'd explode. Now they do.

While here, decided to handle scrolls, spellbooks and globs of green
slime properly as well.