Improvements to showmondamage().
The showmondamage() routine was showing monster damage regardless of the
source. Was causing the message display to become really busy,
espeically if monsters attacked other monsters, or stepped on a trap,
etc. The intent was to only show feedback when the player caused damage
to a monster.

In explode.c it's all set to show TRUE, could use some tweaks. Bit
confusing in that file to determine who's at fault for damage dealt.