Fix: monsters freezing in place at a distance if player was riding a steed that was loyal or greater (tameness >= 15).
Yet another long-standing self-induced bug that I inadvertently
introduced back in August of 2019 via commit e082a28, and continued with
commit f7c1763.

The gist of it is this - if the player was riding a steed, and that
steed was very tame (loyal or fiercely loyal), an attacking monster
would freeze in its tracks if it had any means of attacking at range.
The only way to unfreeze it was to either dismount, make your steed less
tame, or get really close to it while still riding (the last bit,
monster would unfreeze, but still wouldn't necessarily attack). Bizarre
scenario, and this was not the first thing I thought to look at when
troubleshooting this problem. This bit of code had to be completely
reworked.

This fix appears to actually work, I tested several scenarios. A second
set of eyes on this would be welcome though.