Fix: Lucifer could activate the magic portal in the Sanctum.
With the recent changes to how covetous monsters can move, they are now
able to walk normally when close to the player. In a very recent game,
Lucifer walked over the magic portal to Purgatory in the Sanctum to
position himself closer to the player. This in turn activated the
portal, and he went through. The player can't use this portal until
Lucifer has been killed - this set up an unwinnable scenario.

The fix: don't allow Lucifer to activate a magic portal while in the
Sanctum. I put a guard in here for Purgatory/Saint Michael as well. Not
really needed (resists_mgc() monsters can't activate a portal) but
having a guard here anyways isn't a bad idea.