Tweaks to the Druid spells entangle, create grass, create trees.
The biggest change in this commit is how the spell 'entangle' functions.
The affected target must be standing on grass or next to a live tree for
this spell to work. This matches the behavior of the spell from ad&d -
the spell causes existing vegeation to entangle the target. No
vegetation, nothing to entangle with. The entanglement minimum duration
has been adjusted so it lasts a bit longer (the greater the skill in
evocation spells, the greater the entanglement duration, min and max).

To balance this out, the spells 'create grass' and 'create trees' are
now levels 2 and 4 respectively (they were levels 3 and 5 prior to this
commit). This breathes new life into the utility of these spells.