Allow the spell 'create grass' to work in corridors.
Corridors are now fair game as a valid space for grass to grow via druid
magic. I did not allow the spell 'create trees' to work directly in
corridors like this. However, a creative druid could grow grass
throughout the corridors of the dungeon and then create trees on
some of those grassy spots.