New object: scroll of consecration.
This scroll is featured in a handful of variants, such as Fourk and
dNetHack. When read, it creates an altar (unfractured) on the spot the
player is standing on.

Different variants have different behaviors for this scroll, some of
which are the same here. Here's the breakdown:

* A scroll of consecration is a unique object; only one will ever
appear in the game, and it will not appear randomly or in a bones file.
It also cannot be polypiled or wished for
* The quest nemesis for each role has this scroll in their possesion
(inside a sack so it's not out in open inventory)
* When read, the scroll will create an altar on the spot the player is
standing on, and this altar will be based off of the player's alignment.
The altar produced will always be unfractured
* If read while confused, or in Gehennom, or if the scroll is cursed,
the altar produced will always be unaligned (Moloch)
* Restrictions: this scroll will only work on normal tile spaces (room),
sand, or grass. It also cannot be used anywhere on the Planes or in the
Sanctum. If read while standing on a tile space other than a regular
floor/sand/grass, or read while on the Planes or in the Sanctum, nothing
happens, and the scroll is used up

That's it. Fairly simple, and it allows the player to create a
single altar if needed.