Fix: impossible "Can't place branch!" on special levels.
Caught this one (again) while fuzzing. This would be a very rare
occurrence, but still a non-zero chance of it happening, so needed to be
addressed.

On special levels like the nymph level, or the dark forest level, or
even the big room, the quest branch portal can sometimes spawn there. If
there were no free room tile spaces available to place a branch portal,
it would trigger an impossible and cause the game to error out, and
that could potentially lead to an unwinnable scenario (no quest portal,
can't do the quest, etc). This fix makes it so if the game can't place a
branch portal on a room tile, it will pick an open corridor space
instead of triggering an impossible.

This fix also allowed me to get rid of the 'not in gehennom' guard (hell
levels are open mines caverns, no rooms ever).

Having a portal form in a corridor is not ideal, but it's much better
than not forming at all.