Fix: prevent worm tail segment shrink when a long worm is incapacitated.
Found this one whilst fuzzing - when a long worm fails to move,
worm_nomove() is called, which causes the long worm to 'shrink' and lose
a tail segment. This generally happens when a turn passes and the worm
just doesn't move, or can't go where it wants to go because it's
blocked. If failing to move was caused by a condition like being
stunned or trapped (immobile but not really 'stuck'), it could sometimes
cause a desync with wtails/wheads lists, which in turn could cause a
segfault in see_wsegs() (tries to refresh tail segements display, but
they aren't where they're supposed to be).

The fix - guard against worm_nomove() being called when a long worm is
stunned, or trapped, or in an otherwise similar state.