Fix: dmonsfree killing monster using burning hands/shocking grasp while polymorphed.
If using the spells burning hands or shocking grasp, and polymorphed
into a monster that has a claw/touch/kick attack, an extra attack was
snuck into the order of events, and could cause a dmonsfree error if the
target monster was killed via one of these erroneous attacks. This bug
was introduced from commit aea1e41 a few months ago. I used sum[i] here
for the damage instead of just calling the damage directly. Oops.