Attempt to fix macOS build (github workflows) from last commit.
The macOS build was choking on legacy 'no-prototype' declarations from
the latest changes in include/system.h, and on Apple’s Clang they
conflict with the real prototypes in <unistd.h>, causing warnings that
are being treated as errors using the mac build via github workflow.
We'll see if this nips that in the bud.