Fix: virtually any race/role could mount and ride restricted steeds.
Prior to this version (0.9.1) only horses and the like would spawn with
a random saddle (and/or barding, but not relevant to this commit). Now
in 0.9.1, any monster that could potentially be a steed, can randomly
spawn wearing a saddle - this is because monsters can no longer ride
other monsters bareback, they need a saddle.

There are some special cases for what a player can ride: only orcs can
ride wargs, only cavepersons can ride saber-toothed tigers, etc. But the
check to make sure that was enforced was that a player couldn't saddle
(or put barding on) those types of monsters, and that was enough in
previous versions. But now some of these monsters will spawn with a
saddle. A check for the actual act of mounting a steed is now required.

While here, I'm allowing vampires to be able to ride skeletal horses
along with draugr (they're undead).