Monsters will prefer using silver weapons vs demons/vampires.
Monsters will now prefer silver weapons over what would normally be a
stronger weapon if their intended target is a demon or vampire. This
works for both monster vs player and tame monster (pet) vs monster.

Example: you're a vampire [insert role here], and a monster attacking
you has both a katana and a mace, but the mace is made of silver.
Normally that monster would choose the katana, but because you're a
vampire, it will choose the silver mace to attack with instead.

There's probably room for improvement here, but initial testing looked
promising.

There's some unrelated code formatting in this commit as well.