Fix: missing CLC_CALL_UNDEAD case in ucast_cleric_spell().
Found this one while fuzzing. Normally the player is prevented from
accessing this monster spell when polyd into a monster than can cast
clerical spells (no free pets), and there are monster vs monster
restrictions as well. However, when fuzzing the game can sometimes try
to select 'spellnum 15' to cast, and that case for CLC_CALL_UNDEAD was
missing, causing an impossible. Rare, but was a non-zero chance.