Fix: "already have candelabrum?" while fuzzing.
Now that the fuzzer can reach much deeper into the dungeon, it can
realistically obtain the actual invocation objects. But it will still
periodically wish for them as well. And if you have more than one of the
same invocation objects, the game correctly throws an impossible.

The fix - keep the impossible around for regular game play should that
ever occur, but while fuzzing or in wizard mode, kick off a regular
pline message instead warning about the duplicate.