Fix: more "Couldn't place lregion type %d in (%d,%d,%d,%d)!" issues with Gehennom level hella-2.
The issue was that this level uses a mines template, and after the main
part of the fort is created, there's not a lot of tile spaces left over
for stairs placement. The fix is to get rid of the mines fill and do a
percentage of terrain types, just like hella-3 does (decided to have all
of the hella-* maps use this). This opens up many more tile spaces for
stairs placement, and having all of the hella-* levels use this gives a
sense of uniformity with the first tier demon boss levels.