Fix: monsters charging objects inside carried containers.
This started out as chasing an issue from a recent fuzzer session where
a monster had read a scroll of charging, blew up the wand they charged,
and caused m_useup() to throw an impossible about the wand being the
wrong material, which happened to be 0, and caused weight() to divide by
zero. Didn't make any sense at all. And then I recovered the crashed
game and started playing it.

It was an Infidel player monster doing the zapping and reading. And as I
moved, I watched it read a scroll of charging, and it charged up the
wand while it was still in its sack. Which shouldn't be possible. This
condition maybe could set up a bad pointer or corrupt field.

The fix, monsters will intelligently remove the objects they want to
charge out of their carried container first, then charge it. This
properly extracts the object from its bag before its chosen to be
charged and used.