Fix: container weight calculations when a monster removes an object; wand explosion monster/player parity.
Was getting some really weird crashes while fuzzing - exploding wands
not having an assigned material which in turn made it not have any
weight. It was stemming from monsters removing a wand they wanted to
charge from their carried bag, and that wand then blowing up (charged
too many times). The new weight of the container was not being
calculated properly after removal, which appears to be the root cause of
the crashes.

While here, properly refactor exploding wands for monsters, so it has
parity with how it affects the player doing the same, as well as
properly applying damage and handling monster death.