Fix: prevent monster position desynchronization in rloc_to()
Been seeing this crash for a long time, hopefully this fixes it...

The game would panic with "mon at <x,y> is not there!" when rloc_to()
was called to relocate a monster to a square already occupied by another
monster. This caused the monster at the destination to be overwritten in
level.monsters[][] while remaining in the fmon list with its original
coordinates, creating an inconsistent state detected by
mon_sanity_check().

Add a check in rloc_to() to abort relocation if the destination is
occupied, except for the special case of rider/steed pairs which are
allowed to share the same position.