Fix: revert commit 290ad31, proper fix for "No window slots!" crash.
Demo pointed out to me that commit 290ad31 was not an actual fix, but
more a band-aid 'patch the symptom' fix that didn't address the root
cause of the issue. With the inclusion of Druids and vampire race, the
menus I made to allow a player to choose what form they wanted to change
into allowed for a window leak if they had unchanging and couldn't make
a selection. That has now been fixed properly. Thanks Demo.